Refactor Facade and Function to also use the inheritance
