+++ prime-router/src/main/kotlin/reportfilehistory/azure/SubmissionsFacade.kt
@@ -4,12 +4,8 @@ import gov.cdc.prime.router.DetailActionLog
 import gov.cdc.prime.router.DetailReport
 import gov.cdc.prime.router.DetailedSubmissionHistory
 import gov.cdc.prime.router.SubmissionHistory
-import gov.cdc.prime.router.azure.db.tables.pojos.Action
 import gov.cdc.prime.router.common.JacksonMapperUtilities
-import gov.cdc.prime.router.tokens.AuthenticatedClaims
-import org.apache.logging.log4j.kotlin.Logging
 import java.time.OffsetDateTime
-import java.util.UUID
 
 /**
  * Submissions / history API
@@ -17,9 +13,10 @@ import java.util.UUID
  */
 class SubmissionsFacade(
     private val dbSubmissionAccess: SubmissionAccess = DatabaseSubmissionsAccess(),
-    private val dbAccess: DatabaseAccess = WorkflowEngine.databaseAccessSingleton
-) : Logging {
-
+    dbAccess: DatabaseAccess = WorkflowEngine.databaseAccessSingleton
+) : ReportFileFacade(
+    dbAccess,
+) {
     // Ignoring unknown properties because we don't require them. -DK
     private val mapper = JacksonMapperUtilities.allowUnknownsMapper
 
@@ -111,48 +108,6 @@ class SubmissionsFacade(
         submission?.enrichWithSummary()
 
         return submission
-    }
-
-    /**
-     * @return a single Action associated with this [reportId]
-     */
-    fun fetchActionForReportId(reportId: UUID): Action? {
-        return dbAccess.fetchActionForReportId(reportId)
-    }
-
-    /**
-     * @return a single Action, given its key [actionId]
-     */
-    fun fetchAction(actionId: Long): Action? {
-        return dbAccess.fetchAction(actionId)
-    }
-
-    /**
-     * Check whether these [claims] allow access to this [action]
-     * @return true if [claims] authorizes access to this 'receive' [action].  Return
-     * false if the submissionId is not a proper submission or if the claim does not give access.
-     */
-    fun checkSenderAccessAuthorization(
-        action: Action,
-        claims: AuthenticatedClaims,
-    ): Boolean {
-        return when {
-            // Admins always get access
-            claims.isPrimeAdmin -> true
-            // User has an organization claim that matches the action's sendingOrg
-            // No reason to also require that the org have (isSenderOrgClaim == true) because
-            // if they belong to the org that sent it, that's sufficient to say they have the right to see it.
-            (action.sendingOrg == claims.organizationNameClaim) &&
-                (!claims.organizationNameClaim.isNullOrBlank()) -> true
-            else -> {
-                logger.error(
-                    "User from org '${claims.organizationNameClaim}'" +
-                        " denied access to action_id ${action.actionId}" +
-                        " submitted by ${action.sendingOrg}"
-                )
-                false
-            }
-        }
     }
 
     companion object {
